+++ 02 struct/main.go
