+++ 03 pointers/main.go
