+++ 04 pointer2/main.go
