+++ 06 const/main.go
