+++ 10 Interfaces/main.go
@@ -1,0 +1,15 @@
+package main
+
+type speaker interface {
+	speak()
+}
+
+type english struct {
+}
+
+type chinese struct {
+}
+
+func main() {
+
+}
